% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{fixup}
\alias{fixup}
\title{Fixup a data table}
\usage{
fixup(table, overrides)
}
\arguments{
\item{table}{the parameter table to override}

\item{overrides}{the replacement parameters and values}
}
\description{
overrides parameter entries in a table with alternatives from a list
}
