% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{generate_intervention_file}
\alias{generate_intervention_file}
\title{Generate the intervention file}
\usage{
generate_intervention_file(overrides, outpath)
}
\arguments{
\item{overrides, }{the intervention model params that take precedence over the
default parameters}

\item{outpath, }{the path to write the intervention parameters to}
}
\description{
to write a file to parameterise the interventions in the model
}
