% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_overview.R
\name{plot_outputs}
\alias{plot_outputs}
\title{Make some standard plots for the simulation}
\usage{
plot_outputs(OUTPUT)
}
\description{
takes the simulation output and plots standard graphs. For use in
interactive mode only!
}
