% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{present_output}
\alias{present_output}
\title{present output
Adds column names to the model output}
\usage{
present_output(output)
}
\arguments{
\item{output, }{the model output}
}
\description{
present output
Adds column names to the model output
}
