% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_simulation}
\alias{run_simulation}
\title{Run the simulation}
\usage{
run_simulation(
  model = NULL,
  farauti = NULL,
  punctulatus = NULL,
  koliensis = NULL,
  interventions = list(),
  prev_min_ages = 2,
  prev_max_ages = 10,
  incidence_min_ages = c(0, 5, 15),
  incidence_max_ages = c(5, 15, 80),
  a0 = NULL,
  a_seasonality = NULL,
  b_seasonality = NULL
)
}
\arguments{
\item{model, }{A named list of parameters to override in the
"model_parameters.txt" file}

\item{farauti, }{A named list of parameters to override in the
"farauti_parameters.txt" file}

\item{punctulatus, }{A named list of parameters to override in the
"punctulatus_parameters.txt" file}

\item{koliensis, }{A named list of parameters to override in the
"koliensis_parameters.txt" file}

\item{interventions, }{A named list of intervention parameters}

\item{prev_min_ages, }{A vector of minimum ages (in years) to disaggregate prevalence
statistics by.}

\item{prev_max_ages, }{A vector of maximum ages (in years) to disaggregate prevalence
statistics by. Individuals will be summarised by the interval:
[prev_min_ages[i], prev_max_ages[i])}

\item{incidence_min_ages, }{A vector of minimum ages (in years) to disaggregate incidence 
statistics by}

\item{incidence_max_ages, }{A vector of maximum ages (in years) to disaggregate incidence
statistics by. Individuals will be summarised by the interval:
[incidence_min_ages[i], incidence_max_ages[i])}

\item{a0}{If set, the seasonality will be modelled with the fourier method and
a0 will be the first parameter}

\item{a_seasonality}{the a parameters for the fourier model of seasonality}

\item{b_seasonality}{the b parameters for the fourier model of seasonality}
}
\description{
parameterises and runs the model. A table with the simulation
output for each timestep will be returned.

Default parameters can be found in the inst/default/ directory
}
\examples{
output <- run_simulation(model = list(end_time = 1991, bb = 2), farauti=list(mu_M = .5))
dim(output)

}
