% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_simulation}
\alias{run_simulation}
\title{Run the simulation}
\usage{
run_simulation(
  model = NULL,
  farauti = NULL,
  punctulatus = NULL,
  koliensis = NULL
)
}
\arguments{
\item{model}{parameters to override}

\item{farauti}{parameters to override}

\item{punctulatus}{parameters to override}

\item{koliensis}{parameters to override}
}
\description{
parameterises and runs the model. The output will be returned
}
