% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_simulation}
\alias{run_simulation}
\title{Run the simulation}
\usage{
run_simulation(
  model = NULL,
  farauti = NULL,
  punctulatus = NULL,
  koliensis = NULL
)
}
\arguments{
\item{model, }{A named list of parameters to override in the
"model_parameters.txt" file}

\item{farauti, }{A named list of parameters to override in the
"farauti_parameters.txt" file}

\item{punctulatus, }{A named list of parameters to override in the
"punctulatus_parameters.txt" file}

\item{koliensis, }{A named list of parameters to override in the
"koliensis_parameters.txt" file}
}
\description{
parameterises and runs the model. A table with the simulation
output for each timestep will be returned.

Default parameters can be found in the inst/default/ directory
}
\examples{
output <- run_simulation(model = list(end_time = 1991, bb = 2), farauti=list(mu_M = .5))
dim(output)

}
