% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_simulation_from_path}
\alias{run_simulation_from_path}
\title{Entrypoint for the simulation}
\usage{
run_simulation_from_path(
  model_param_path,
  fara_param_path,
  punc_param_path,
  koli_param_path,
  coverage_param_path,
  output_path
)
}
\arguments{
\item{model_param_path, }{the path of the model parameter file}

\item{fara_param_path, }{the path to the farauti mosquito parameter files.}

\item{punc_param_path, }{the path to the punctulatus mosquito parameter files.}

\item{koli_param_path, }{the path to the koliensis mosquito parameter files.}

\item{coverage_param_path, }{the path of the coverage parameter file}

\item{output_path, }{the path to write the results to}
}
\description{
Entrypoint for the simulation
}
